[NEVER UPSTREAM] arm64: dts: uniphier: add simple-framebuffer and its relate
Add simple-framebuffer node to Akebi96, which is provided by boot
firmware.

It is necessary to allocate and set physical address and size for
framebuffer that is configurated by firmware. And the area indicated by
the address and size must be also defined by reserved-memory node with
'no-map' because simplefb driver remaps this area.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>